+++ display/X/pixel.c
@@ -1,4 +1,4 @@
-/* $Id: pixel.c,v 1.2 2002/09/08 21:37:44 soyt Exp $
+/* $Id: pixel.c,v 1.3 2003/07/06 10:25:21 cegger Exp $
 ******************************************************************************
 
    Graphics library for GGI.  Pixels for display-X.
@@ -212,7 +212,7 @@ int GGI_X_getpixel_draw(ggi_visual *vis, int x, int y, ggi_pixel *pixel)
 	noswab:
 		ximgptr = ximg->data + 
 			(ximg->xoffset * ximg->bits_per_pixel)/8;
-		memcpy(pixel, ximgptr, ximg->bits_per_pixel/8);
+		memcpy(pixel, ximgptr, (size_t)ximg->bits_per_pixel/8);
 	}
 	XDestroyImage(ximg);
  out:
